+++ src/http/extractHttpUserFromHeaders.ts
